Refactor harness: remove redfield/greenfield, make brownfield default, and split checks by feature
- Removed `greenfield` and `redfield` scenarios from config and app checks.
- Set `brownfield` as the default and only scenario.
- Extracted speculation rules checks from `brownfield.ts` into `speculation_rules.ts`.
- Updated metrics calculation and tests to reflect the removal of scenarios.
